requirements.txt
flake8
h5py
jupyter
keras
matplotlib
pillow
sklearn
tensorflow
